move jobs to different file, so they can be imported without triggering job loop
